mass db transactions & date fetch fix

+++ database/migrations/2025_02_08_102911_create_moviews_table.php
@@ -25,13 +25,13 @@ return new class extends Migration
 
         Schema::create('budgets', function (Blueprint $table) {
             $table->id();
-            $table->integer('budget');
+            $table->decimal('budget', 10, 2);
             $table->timestamps();
         });
 
         Schema::create('box_offices', function (Blueprint $table) {
             $table->id();
-            $table->integer('revenue');
+            $table->decimal('revenue', 10, 2);
             $table->timestamps();
         });
